+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.2 (13/09/2019)
+* __Enhancement__: Redeveloped the plugin to include front matter for the config
+* __Enhancement__: Added commands to load, save, clear, and delete the config
+
 ## 0.0.1 (11/09/2019)
 * Initial release